+++ core/app/(user)/signup/page.tsx
@@ -1,6 +1,6 @@
 import { AuthTokenType } from "db/public";
 
-import { SignupForm } from "~/app/components/Signup/BaseSignupForm";
+import { SignupForm } from "~/app/components/Signup/SignupForm";
 import { legacySignup } from "~/lib/authentication/actions";
 import { getLoginData } from "~/lib/authentication/loginData";
